requirements.txt
torch>=2.1
transformers>=4.41
datasets>=2.19
accelerate>=0.31
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
biopython>=1.83
matplotlib>=3.7
